move struct CardDataC to data_manager.h
