+++ gframe/deck_manager.cpp
@@ -1,6 +1,5 @@
 #include "deck_manager.h"
 #include "myfilesystem.h"
-#include "data_manager.h"
 #include "network.h"
 #include "game.h"
