+++ gframe/single_duel.h
@@ -1,6 +1,7 @@
 #ifndef SINGLE_DUEL_H
 #define SINGLE_DUEL_H
 
+#include <set>
 #include "network.h"
 #include "deck_manager.h"
 #include "replay.h"
